Update README.md with a note on Docker memory
Most dev team members seem to allocate ~6GiB to Docker. This should be documented as part of the setup process because it's painful to see things lag & crash with the default 1GiB on Mac.

Anecdotally, increasing memory was necessary for me to successfully run the './project.rb dev-up' command in the workbench/ui folder.