Work in progress on adding Spring GCP trace module.
Kind of a rabbit hole here... it looks like we'll get better
Stackdriver Trace visibility if we use the Spring GCP Trace
module. But that module isn't compatible with our older version of
Spring, so this WIP commit is an attempt to upgrade the whole thing in
one go, just to see what breaks. It looks like a few APIs are
deprecated / changed, and some Tomcat stuff is failing.

Putting this on the backburner; I'll circle back at some point to take
another look.