[risk=no] Fix firecloud.yaml typo
I don't know how this happened - the source api-docs.yaml is correct (and has been for at least 3y), but it cost me ~1h of debugging to find. Also some weirdness around `workbenchRole` vs `role`; I don't know whether that was originally an issue in the docs or that was a problem of our own design - this endpoint never worked and the only place we call this, we're suppressing exceptions.