fixes for VS Code editor (#2442)
When editing/viewing the UI code in Visual Studio Code, I noticed that paths weren't resolving correctly. It looks like this is because we're using webpack implied roots (like app). Adding paths for app and generated to .tsconfig fixed that.

Additionally, I want to ignore the .vscode directory this editor creates.

An open question is whether this can go in a shared .editorconfig.