Remove the section “How to Use the All of Us Researcher Workbench” from home page.
Styles and HTML were removed. Margin was added to the bottom between the main page and the bottom banner.

The goal is to re-add this in the future. I opted to delete the code rather than commenting the code out in order to keep the code base a bit cleaner. We can pull in whatever we need at that point.