Validate and auto-update state field to state code
If the institutional address is within the US, validates that
the state is a 2-letter state code (CA, TX, etc.).

When the user selects US, auto-update the state field to a code
if we can detect the state they've entered.

This should clean a lot of the state data going forward. Before this,
users were entering a wide variety of state values.