[RW-8731][risk=no] Validate and clean country and state address data (#7006)
* Sanitize mailing address country input

Changes the 'country' field of the institutional mailing address from
an open text input to a dropdown containing "United States of America"
and "Other".

This should reduce the amount of unclean data on that field. Users
were entering many variations of USA and occasionally county info.

* Validate and auto-update state field to state code

If the institutional address is within the US, validates that
the state is a 2-letter state code (CA, TX, etc.).

When the user selects US, auto-update the state field to a code
if we can detect the state they've entered.

This should clean a lot of the state data going forward. Before this,
users were entering a wide variety of state values.