Fix: Update broken fontawesome data-icon path
"sync-alt" is an alias for the "rotate" icon in fontawesome. I suspect
this is a breaking change from the fontawesome 6.0 upgrade. We applied
that update in #6895 at the end of July, about a week after this test
starting failing due to another issue, which is why we didn't detect
this sooner.