Combine controlled tier test with happy-path test
I think the core usage of this module is using it
once for RT then once for CT, so that should be
reflected in the basic test.

Otherwise, the controlled tier is treated the same
as the RT, so there's no need for a renewal test
specifically for CT. Those checks are moved into
the renewal test.

This allows us to remove the controlled tier test.